# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fingerprint Time Logger API
  description: |
    Comprehensive API for managing ZKTeco biometric devices, employee records, and attendance tracking.
    
    ## Features
    - **ZKTeco Device Integration**: Direct fingerprint device communication
    - **Employee Management**: Unified employee records with Thai/English names
    - **Attendance Tracking**: Real-time punch data collection and display
    - **System Monitoring**: Health checks, diagnostics, and performance metrics
    - **Data Export**: CSV export with flexible filtering options
    - **Background Tasks**: Automated data synchronization
    
    ## Authentication
    Currently, the API does not require authentication. It's designed for single-user local installations.
    
    ## Real-time Updates
    The system supports WebSocket connections for real-time updates at `ws://localhost:5000/ws`.
    
    ## Error Handling
    All endpoints return consistent error responses with appropriate HTTP status codes.
    
  version: 2.0.0
  contact:
    name: API Support
    url: https://github.com/your-org/fingerprint-time-logger
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://your-domain.com
    description: Production server
...
components:
  schemas:
...
    Device:
      type: object
      properties:
        id:
          type: integer
          description: Unique device identifier
        name:
          type: string
          description: Device display name
        ip_address:
          type: string
          format: ipv4
          description: Device IP address
        port:
          type: integer
          default: 4370
          description: Device port number
        password:
          type: integer
          default: 0
          description: Device password
        is_active:
          type: boolean
          default: true
          description: Device active status
        last_sync:
          type: string
          format: date-time
          nullable: true
          description: Last synchronization timestamp
        created_at:
          type: string
          format: date-time
          description: Device creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Device last update timestamp
      required:
        - id
        - name
        - ip_address
        - port
        - is_active
        - created_at
        - updated_at
    Employee:
      type: object
      properties:
        id:
          type: integer
          description: Unique employee identifier
        badge_number:
          type: string
          description: Employee badge number (unique identifier)
        english_name:
          type: string
          nullable: true
          description: Employee English name
        thai_name:
          type: string
          nullable: true
          description: Employee Thai name
        display_name:
          type: string
          description: Employee display name (computed from Thai name or badge)
        department:
          type: string
          nullable: true
          description: Employee department
        position:
          type: string
          nullable: true
          description: Employee position
        job_role_id:
          type: integer
          nullable: true
          description: Associated job role ID
        is_active:
          type: boolean
          default: true
          description: Employee active status
        is_hidden:
          type: boolean
          default: false
          description: Employee visibility status
        created_at:
          type: string
          format: date-time
          description: Employee creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Employee last update timestamp
      required:
        - id
        - badge_number
        - display_name
        - is_active
        - is_hidden
        - created_at
        - updated_at
    AttendanceRecord:
      type: object
      properties:
        id:
          type: integer
          description: Unique attendance record identifier
        employee_badge_number:
          type: string
          description: Employee badge number
        device_id:
          type: integer
          description: Device ID that recorded the punch
        timestamp:
          type: string
          format: date-time
          description: Punch timestamp
        punch_type:
          type: integer
          enum: [0, 1, 2, 3, 4, 5]
          description: |
            Punch type:
            - 0: Check In
            - 1: Check Out
            - 2: Break Out
            - 3: Break In
            - 4: Overtime In
            - 5: Overtime Out
        status:
          type: integer
          enum: [0, 1, 2]
          default: 0
          description: |
            Punch status:
            - 0: Normal
            - 1: Late
            - 2: Early
        sync_status:
          type: string
          enum: [synced, pending, failed]
          default: synced
          description: Synchronization status
        created_at:
          type: string
          format: date-time
          description: Record creation timestamp
        employee:
          $ref: '#/components/schemas/Employee'
        device:
          $ref: '#/components/schemas/Device'
      required:
        - id
        - employee_badge_number
        - device_id
        - timestamp
        - punch_type
        - status
        - sync_status
        - created_at
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
          description: Operation success status
        error:
          type: string
          description: Error message
        detail:
          type: string
          nullable: true
          description: Detailed error information
        code:
          type: string
          nullable: true
          description: Error code
      required:
        - success
        - error
...
    AttendanceResponse:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/AttendanceRecord'
        total:
          type: integer
          description: Total number of records
        filters:
          type: object
          properties:
            start_date:
              type: string
              format: date-time
              nullable: true
            end_date:
              type: string
              format: date-time
              nullable: true
            employee_badge:
              type: string
              nullable: true
      required:
        - records
        - total
        - filters
    DeviceStatus:
      type: object
      properties:
        device_id:
          type: integer
          description: Device identifier
        connected:
          type: boolean
          description: Connection status
        ip_address:
          type: string
          format: ipv4
          description: Device IP address
        port:
          type: integer
          description: Device port
        firmware:
          type: string
          nullable: true
          description: Device firmware version
        users_count:
          type: integer
          nullable: true
          description: Number of users on device
        records_count:
          type: integer
          nullable: true
          description: Number of records on device
        last_connection:
          type: string
          format: date-time
          nullable: true
          description: Last successful connection
        response_time_ms:
          type: integer
          nullable: true
          description: Response time in milliseconds
      required:
        - device_id
        - connected
        - ip_address
        - port
...
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /api/attendance/:
    get:
      tags:
        - Attendance
      summary: Get attendance records
      description: Retrieve attendance records with optional filtering and pagination
      parameters:
        - name: start_date
          in: query
          description: Filter start date (ISO format)
          required: false
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          description: Filter end date (ISO format)
          required: false
          schema:
            type: string
            format: date-time
        - name: employee_badge
          in: query
          description: Filter by employee badge number
          required: false
          schema:
            type: string
        - name: limit
          in: query
          description: Number of records to return
          required: false
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          description: Record offset for pagination
          required: false
          schema:
            type: integer
            default: 0
      responses:
        200:
          description: Attendance records retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /api/devices/:
    get:
      tags:
        - Devices
      summary: Get all devices
      description: Retrieve all registered ZKTeco devices
      responses:
        200:
          description: Devices retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /api/devices/status:
    get:
      tags:
        - Devices
      summary: Get device status
      description: Get current device connection status and information
      responses:
        200:
          description: Device status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceStatus'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /api/employees/:
    get:
      tags:
        - Employees
      summary: Get all employees
      description: Retrieve all employees with optional filtering
      parameters:
        - name: include_inactive
          in: query
          description: Include inactive employees
          required: false
          schema:
            type: boolean
            default: false
        - name: include_hidden
          in: query
          description: Include hidden employees
          required: false
          schema:
            type: boolean
            default: false
        - name: role_id
          in: query
          description: Filter by job role ID
          required: false
          schema:
            type: integer
      responses:
        200:
          description: Employees retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  employees:
                    type: array
                    items:
                      $ref: '#/components/schemas/Employee'
                  total:
                    type: integer
                  active:
                    type: integer
                  hidden:
                    type: integer
        500:
          $ref: '#/components/responses/InternalServerError'
...